consolidate fixtures, fix some tests
