--- test/fixture/document-types-response.json
@@ -1,17 +1,0 @@
-[
-  {
-    "id": "passport-scan",
-    "title": "Passport",
-    "description": "A valid passport open at the photo page"
-  },
-  {
-    "id": "driving-license",
-    "title": "Driving license",
-    "description": "A valid UK full or provisional UK driving license"
-  },
-  {
-    "id": "bank-statement",
-    "title": "Bank statement",
-    "description": "From the past three months"
-  }
-]
