--- test/fixture/evidence-request-response.json
@@ -1,53 +1,0 @@
-[
-  {
-    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-    "createdAt": "2020-11-30T15:34:12.299Z",
-    "resident": {
-      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-      "name": "Namey McName",
-      "email": "frodo@bagend.com",
-      "phoneNumber": "+447123456780"
-    },
-    "deliveryMethods": ["SMS", "EMAIL"],
-    "documentType": {
-      "id": "passport-scan",
-      "title": "Passport",
-      "description": "A valid passport open at the photo page"
-    },
-    "serviceRequestedBy": "development-team-staging"
-  },
-  {
-    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-    "createdAt": "2020-11-29T15:34:12.299Z",
-    "resident": {
-      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-      "name": "Namey McName",
-      "email": "frodo@bagend.com",
-      "phoneNumber": "+447123456780"
-    },
-    "deliveryMethods": ["SMS", "EMAIL"],
-    "documentType": {
-      "id": "passport-scan",
-      "title": "Passport",
-      "description": "A valid passport open at the photo page"
-    },
-    "serviceRequestedBy": "development-team-staging"
-  },
-  {
-    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-    "createdAt": "2020-11-25T15:34:12.299Z",
-    "resident": {
-      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-      "name": "Namey McName",
-      "email": "frodo@bagend.com",
-      "phoneNumber": "+447123456780"
-    },
-    "deliveryMethods": ["SMS", "EMAIL"],
-    "documentType": {
-      "id": "passport-scan",
-      "title": "Passport",
-      "description": "A valid passport open at the photo page"
-    },
-    "serviceRequestedBy": "development-team-staging"
-  }
-]
